+++ CHANGES.rst
@@ -9,6 +9,7 @@ New Features
 * #35: Added ``BaseModel.dump_to_csv()``
 * #35: Added ``BaseModel.dump_to_json()``
 * #35: Added ``BaseModel.dump_to_yaml()``
+* #35: Added ``BaseModel.dump_to_dict()``
 
 -----------